Give each value list its own file and icon
