+++ grasshopper-plugin/TapirGrasshopperPlugin/Properties/Resources.resx
@@ -895,4 +895,31 @@
   <data name="GetMEPPreferenceTables" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\GetMEPPreferenceTables.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
+  <data name="StructureTypeValueList" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\StructureTypeValueList.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="ReferenceLineLocationValueList" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\ReferenceLineLocationValueList.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="ReferencePlaneLocationValueList" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\ReferencePlaneLocationValueList.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="AnchorPointValueList" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\AnchorPointValueList.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="TextJustificationValueList" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\TextJustificationValueList.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="MeshSkirtTypeValueList" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\MeshSkirtTypeValueList.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="MeshRidgeTypeValueList" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\MeshRidgeTypeValueList.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="DimensioningPresetValueList" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\DimensioningPresetValueList.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="MEPDomainValueList" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\MEPDomainValueList.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
 </root>
